fixed pie chart for relational data (key/groupBy is on another table)
This resolves the bug which prevented the use of the pie chart with relational data. Please be aware that field.type will always be undefined as it only checks fields on the originating table. 
Row 11 would need to be adjusted to check all associations, too. But I didn't know how to properly implement that, sorry. So date formatting is not being done in this case. But besides that little thing, it works.