Sequelize 4 returns date strings not timestamps so don't shift
As of Sequelize 4 dateonly fields are returned as date strings not
dateTime strings.  Therefore the formatting functionality here will
return incorrect values for dateonly fields in Sequelize 4 if the server
is in a timezeon with a negative UTC shift.

Ex: I have 1990-10-02 stored in my DB.  The formatting function will
shift that to 1990-10-01T16:00:00 if I am in PST.  In the UI 01/10/1990
will be shown

sequelize/sequelize#4858